fix WinningPlayerIds as reported by alior in discord

+++ src/FortniteReplayReader/Models/GameData.cs
@@ -24,6 +24,6 @@ namespace FortniteReplayReader.Models
         public float? SafeZonesStartTime { get; set; }
 
         public uint? WinningTeam { get; set; }
-        public IEnumerable<byte> WinningPlayerIds { get; set; }
+        public IEnumerable<int> WinningPlayerIds { get; set; }
     }
 }
